+++ lib/pact_broker/domain/label.rb
@@ -3,7 +3,6 @@ require 'pact_broker/db'
 module PactBroker
   module Domain
     class Label < Sequel::Model
-
       unrestrict_primary_key
 
       associate(:many_to_one, :pacticipant, :class => "PactBroker::Domain::Pacticipant", :key => :pacticipant_id, :primary_key => :id)