+++ public/javascripts/matrix.js
@@ -22,18 +22,26 @@ function showApplicableTextBoxes(selectorizor) {
   var selectorizorType = selectorizor.val();
   if( selectorizorType === 'specify-version') {
     setTextboxVisibility(selectorizor, '.version', true);
+    setTextboxVisibility(selectorizor, '.branch', false);
     setTextboxVisibility(selectorizor, '.tag', false);
   }
   else if( selectorizorType === 'specify-latest-tag' || selectorizorType === 'specify-all-tagged') {
     setTextboxVisibility(selectorizor, '.version', false);
+    setTextboxVisibility(selectorizor, '.branch', false);
     setTextboxVisibility(selectorizor, '.tag', true);
+  }
+  else if( selectorizorType === 'specify-latest-branch') {
+    setTextboxVisibility(selectorizor, '.version', false);
+    setTextboxVisibility(selectorizor, '.branch', true);
+    setTextboxVisibility(selectorizor, '.tag', false);
   }
   else if ( selectorizorType === 'specify-all-versions' || selectorizorType === 'specify-latest') {
     setTextboxVisibility(selectorizor, '.version', false);
+    setTextboxVisibility(selectorizor, '.branch', false);
     setTextboxVisibility(selectorizor, '.tag', false);
   }
 
-  if (selectorizorType === 'specify-latest' || selectorizorType === 'specify-latest-tag') {
+  if (selectorizorType === 'specify-latest' || selectorizorType === 'specify-latest-tag' || selectorizorType === 'specify-latest-branch') {
     toggleLatestFlag(selectorizor, true);
   } else {
     toggleLatestFlag(selectorizor, false);