+++ spec/lib/pact_broker/api/decorators/pact_version_decorator_spec.rb
@@ -27,7 +27,7 @@ module PactBroker
           name: 'pact_name')}
         let(:consumer) { instance_double(PactBroker::Domain::Pacticipant, name: 'Consumer')}
         let(:provider) { instance_double(PactBroker::Domain::Pacticipant, name: 'Provider')}
-        let(:consumer_version) { instance_double(PactBroker::Domain::Version, number: '1234', pacticipant: consumer)}
+        let(:consumer_version) { instance_double(PactBroker::Domain::Version, number: '1234', branch: 'main', pacticipant: consumer)}
         let(:decorator_context) { DecoratorContext.new(base_url, '', {}) }
 
         let(:json) { PactVersionDecorator.new(pact).to_json(user_options: decorator_context) }